requirements.txt
Django~=5.2.4
python-dotenv~=1.1.1
djangorestframework>=3.15.2
django-filter>=24.2
psycopg2>=2.9.9
python-decouple>=3.8
django-environ>=0.11.2

requests~=2.32.4